requirements.txt
asgiref==3.11.0
coverage==7.13.2
Django==6.0.1
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.5
